+++ packages/payments/src/index.ts
@@ -1,0 +1,8 @@
+// CHANGED: exports
+
+// Customization options and base urls
+export * from "./constants";
+// Types
+export * from "./interfaces";
+// Components
+export * from "./lib";