+++ packages/react/src/index.ts
@@ -113,3 +113,4 @@ export type {
 
 // at the moment we'll re-export everything from the evm package
 export * from "./evm";
+export * from "./payments";