+++ packages/react/src/payments/additional.d.ts
@@ -1,0 +1,8 @@
+// UNCHANGED
+declare module '*.png';
+declare module '*.webp';
+declare module '*.jpg';
+declare module '*.jpeg';
+declare module '*.svg';
+declare module '*.gif';
+declare module '*.css';